// src/app/models/user.model.ts
export interface User{
    firstName: string;
    lastName: string;
    email: string;
    mobile: number;
    weight: number;
    height: number;
    bmi:number;
    bmiResult: string;
    gender: string;
    requireTrainer: string;
    package: string;
    important: string[];
    haveGymBefore: string;
    enquiryDate: string;
    id:number;
 
}